Add IgnoreException decorators
The default behavior of the vscode debugger is to only break at uncaught exceptions, but this creates problems when debugging templates, as it causes me to lose my place in the call stack where the error actually occurred.

So, in order to debug templates more effectively, I have ticked the "Raised Exceptions" checkbox in the breakpoint section of the vscode debugger, which makes the debugger break ANY TIME an exception is raised, regardless of whether it is caught or uncaught.

However, there are some caught exceptions which happen quite often and which are of no interest to me, so for these I am adding an IgnoreException decorator.